refactor(ai): split sendMessage into entry + internals + types (559→209)
sendMessage.ts was 559 LOC: the sendAiMessage orchestrator plus a wall of 10
internal helpers and 4 types. Extracted (verbatim, behavior-preserving):
- sendMessage-types.ts — AIProvider/AssistantRecord/SendMessageResult/Error
  (own module so entry + internals share them without a circular import)
- sendMessage-internals.ts — verifyConversation, fetchAssistant, resolveProvider,
  selectModel, checkFreeMessageUsage, buildMessageHistory, callAi,
  isAiRateLimitError, storeUserMessage, storeAssistantMessage

sendMessage.ts (209 LOC) is now just the sendAiMessage pipeline; it re-exports the
two result types so the public API (the messages route) is unchanged.

Note: this isAiRateLimitError intentionally differs from cat/chat-orchestrator's
(also matches status===429 + a regex) — kept separate to preserve each path's
exact behavior rather than force a merge.

tsc 0 errors.
