+++ src/services/ai/sendMessage-types.ts
@@ -1,0 +1,56 @@
+/**
+ * Shared types for the AI assistant message-send pipeline (sendMessage.ts +
+ * sendMessage-internals.ts). Kept in their own module so the entry point and the
+ * internal helpers can both import them without a circular dependency.
+ */
+
+export type AIProvider = 'openrouter' | 'groq';
+
+export interface AssistantRecord {
+  id: string;
+  title: string;
+  system_prompt: string | null;
+  welcome_message: string | null;
+  pricing_model: string;
+  price_per_message: number | null;
+  price_per_1k_tokens: number | null;
+  user_id: string;
+  model_preference: string | null;
+  allowed_models: string[] | null;
+  min_model_tier: string | null;
+  temperature: number | null;
+  max_tokens_per_response: number | null;
+  free_messages_per_day: number | null;
+}
+
+export interface SendMessageResult {
+  userMessage: Record<string, unknown>;
+  assistantMessage: Record<string, unknown>;
+  usage: {
+    inputTokens: number;
+    outputTokens: number;
+    totalTokens: number;
+    apiCostBtc: number;
+    totalCostBtc: number;
+  };
+  userStatus: {
+    hasByok: boolean;
+    usedFreeMessage: boolean;
+    freeMessagesRemaining: number;
+    freeMessagesPerDay: number;
+  };
+}
+
+export type SendMessageError =
+  | { code: 'NOT_FOUND'; message: string }
+  | { code: 'ARCHIVED' }
+  | { code: 'RATE_LIMITED'; message: string }
+  | { code: 'SERVICE_UNAVAILABLE' }
+  | {
+      code: 'INSUFFICIENT_CREDITS';
+      currentBalance: number;
+      requiredAmount: number;
+      shortfall: number;
+    }
+  | { code: 'AI_ERROR'; message: string }
+  | { code: 'DB_ERROR'; message: string };
